# bin/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Api Documentation",
    "version": "1.0",
    "title": "Api Documentation",
    "termsOfService": "urn:tos",
    "contact": {},
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
  "tags": [
    {
      "name": "basic-error-controller",
      "description": "Basic Error Controller"
    },
    {
      "name": "kullanici-servis-controller",
      "description": "Kullanici Servis Controller"
    },
    {
      "name": "tbs-servis-controller",
      "description": "TBS Servis Controller"
    }
  ],
  "paths": {
    "/error": {
      "get": {
        "tags": [
          "basic-error-controller"
        ],
        "summary": "error",
        "operationId": "errorUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "head": {
        "tags": [
          "basic-error-controller"
        ],
        "summary": "error",
        "operationId": "errorUsingHEAD",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "basic-error-controller"
        ],
        "summary": "error",
        "operationId": "errorUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "put": {
        "tags": [
          "basic-error-controller"
        ],
        "summary": "error",
        "operationId": "errorUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "basic-error-controller"
        ],
        "summary": "error",
        "operationId": "errorUsingDELETE",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      },
      "options": {
        "tags": [
          "basic-error-controller"
        ],
        "summary": "error",
        "operationId": "errorUsingOPTIONS",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      },
      "patch": {
        "tags": [
          "basic-error-controller"
        ],
        "summary": "error",
        "operationId": "errorUsingPATCH",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      }
    },
    "/login": {
      "post": {
        "tags": [
          "kullanici-servis-controller"
        ],
        "summary": "verifyUserLDAP",
        "operationId": "verifyUserLDAPUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "projectEnterVerifyLDAP",
            "description": "projectEnterVerifyLDAP",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProjeGirisDogrulaLDAPRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/KullaniciYetkiResponse"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/plakadan-arac-durumu-getir": {
      "post": {
        "tags": [
          "tbs-servis-controller"
        ],
        "summary": "plakadanAracDurumuGetir",
        "operationId": "plakadanAracDurumuGetirUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "aracPlakaRequest",
            "description": "aracPlakaRequest",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TBSAracPlakaRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/AracDurumlariResponse"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/plakadan-arac-sahibi-getir": {
      "post": {
        "tags": [
          "tbs-servis-controller"
        ],
        "summary": "plakadanAracSahibiGetir",
        "operationId": "plakadanAracSahibiGetirUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "aracPlakaRequest",
            "description": "aracPlakaRequest",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TBSAracPlakaRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/AracSahibiResponse"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/project-user-list": {
      "post": {
        "tags": [
          "kullanici-servis-controller"
        ],
        "summary": "getProjectList",
        "operationId": "getProjectListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "auth",
            "description": "auth",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProjectUserList"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/KullaniciYetkiResponse"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    }
  },
  "definitions": {
    "AracDurumlariResponse": {
      "type": "object",
      "properties": {
        "aracDurumu": {
          "$ref": "#/definitions/AracDurumuDTOResponse"
        },
        "hataKodu": {
          "type": "string"
        },
        "hatadetay": {
          "type": "string"
        }
      },
      "title": "AracDurumlariResponse"
    },
    "AracDurumuDTOResponse": {
      "type": "object",
      "properties": {
        "calintiDurumu": {
          "type": "string"
        },
        "plakaDurumu": {
          "type": "string"
        }
      },
      "title": "AracDurumuDTOResponse"
    },
    "AracSahibiResponse": {
      "type": "object",
      "properties": {
        "gercekSahis": {
          "$ref": "#/definitions/GercekSahisDTOResponse"
        },
        "hataKodu": {
          "type": "string"
        },
        "hatadetay": {
          "type": "string"
        },
        "tuzelSahis": {
          "$ref": "#/definitions/TuzelSahisDTOResponse"
        }
      },
      "title": "AracSahibiResponse"
    },
    "GercekSahisDTOResponse": {
      "type": "object",
      "properties": {
        "ad": {
          "type": "string"
        },
        "ad2": {
          "type": "string"
        },
        "dogumTarihi": {
          "type": "string"
        },
        "dogumYeri": {
          "type": "string"
        },
        "mernisNo": {
          "type": "string"
        },
        "soyad": {
          "type": "string"
        }
      },
      "title": "GercekSahisDTOResponse"
    },
    "KullaniciBilgi": {
      "type": "object",
      "properties": {
        "ip": {
          "type": "string"
        },
        "reasonCode": {
          "type": "string"
        },
        "reasonText": {
          "type": "string"
        },
        "tckn": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      },
      "title": "KullaniciBilgi"
    },
    "KullaniciYetki": {
      "type": "object",
      "properties": {
        "mADSOYAD": {
          "type": "string"
        },
        "mBIRLIKKOD": {
          "type": "string"
        },
        "mDAIRE": {
          "type": "string"
        },
        "mEPOSTA": {
          "type": "string"
        },
        "mGOREV": {
          "type": "string"
        },
        "mHATASAYISI": {
          "type": "string"
        },
        "mIL": {
          "type": "string"
        },
        "mILCE": {
          "type": "string"
        },
        "mIP": {
          "type": "string"
        },
        "mKULLANICIADI": {
          "type": "string"
        },
        "mKULLANIMDURUMU": {
          "type": "string"
        },
        "mMASK": {
          "type": "string"
        },
        "mPBIK": {
          "type": "string"
        },
        "mPROJEADI": {
          "type": "string"
        },
        "mPROJEGRUPADI": {
          "type": "string"
        },
        "mPROJEROLKODU": {
          "type": "string"
        },
        "mRUTBE": {
          "type": "string"
        },
        "mSUBE": {
          "type": "string"
        },
        "mTCKMLNMR": {
          "type": "string"
        }
      },
      "title": "KullaniciYetki"
    },
    "KullaniciYetkiResponse": {
      "type": "object",
      "properties": {
        "mHATADETAY": {
          "type": "string"
        },
        "mHATAKOD": {
          "type": "integer",
          "format": "int32"
        },
        "mKullaniciYetkiListe": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/KullaniciYetki"
          }
        }
      },
      "title": "KullaniciYetkiResponse"
    },
    "ModelAndView": {
      "type": "object",
      "properties": {
        "empty": {
          "type": "boolean"
        },
        "model": {
          "type": "object"
        },
        "modelMap": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        },
        "reference": {
          "type": "boolean"
        },
        "status": {
          "type": "string",
          "enum": [
            "100 CONTINUE",
            "101 SWITCHING_PROTOCOLS",
            "102 PROCESSING",
            "103 CHECKPOINT",
            "200 OK",
            "201 CREATED",
            "202 ACCEPTED",
            "203 NON_AUTHORITATIVE_INFORMATION",
            "204 NO_CONTENT",
            "205 RESET_CONTENT",
            "206 PARTIAL_CONTENT",
            "207 MULTI_STATUS",
            "208 ALREADY_REPORTED",
            "226 IM_USED",
            "300 MULTIPLE_CHOICES",
            "301 MOVED_PERMANENTLY",
            "302 FOUND",
            "302 MOVED_TEMPORARILY",
            "303 SEE_OTHER",
            "304 NOT_MODIFIED",
            "305 USE_PROXY",
            "307 TEMPORARY_REDIRECT",
            "308 PERMANENT_REDIRECT",
            "400 BAD_REQUEST",
            "401 UNAUTHORIZED",
            "402 PAYMENT_REQUIRED",
            "403 FORBIDDEN",
            "404 NOT_FOUND",
            "405 METHOD_NOT_ALLOWED",
            "406 NOT_ACCEPTABLE",
            "407 PROXY_AUTHENTICATION_REQUIRED",
            "408 REQUEST_TIMEOUT",
            "409 CONFLICT",
            "410 GONE",
            "411 LENGTH_REQUIRED",
            "412 PRECONDITION_FAILED",
            "413 PAYLOAD_TOO_LARGE",
            "413 REQUEST_ENTITY_TOO_LARGE",
            "414 URI_TOO_LONG",
            "414 REQUEST_URI_TOO_LONG",
            "415 UNSUPPORTED_MEDIA_TYPE",
            "416 REQUESTED_RANGE_NOT_SATISFIABLE",
            "417 EXPECTATION_FAILED",
            "418 I_AM_A_TEAPOT",
            "419 INSUFFICIENT_SPACE_ON_RESOURCE",
            "420 METHOD_FAILURE",
            "421 DESTINATION_LOCKED",
            "422 UNPROCESSABLE_ENTITY",
            "423 LOCKED",
            "424 FAILED_DEPENDENCY",
            "425 TOO_EARLY",
            "426 UPGRADE_REQUIRED",
            "428 PRECONDITION_REQUIRED",
            "429 TOO_MANY_REQUESTS",
            "431 REQUEST_HEADER_FIELDS_TOO_LARGE",
            "451 UNAVAILABLE_FOR_LEGAL_REASONS",
            "500 INTERNAL_SERVER_ERROR",
            "501 NOT_IMPLEMENTED",
            "502 BAD_GATEWAY",
            "503 SERVICE_UNAVAILABLE",
            "504 GATEWAY_TIMEOUT",
            "505 HTTP_VERSION_NOT_SUPPORTED",
            "506 VARIANT_ALSO_NEGOTIATES",
            "507 INSUFFICIENT_STORAGE",
            "508 LOOP_DETECTED",
            "509 BANDWIDTH_LIMIT_EXCEEDED",
            "510 NOT_EXTENDED",
            "511 NETWORK_AUTHENTICATION_REQUIRED"
          ]
        },
        "view": {
          "$ref": "#/definitions/View"
        },
        "viewName": {
          "type": "string"
        }
      },
      "title": "ModelAndView"
    },
    "ProjeGirisDogrulaLDAPRequest": {
      "type": "object",
      "properties": {
        "ip": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "pbik": {
          "type": "string"
        }
      },
      "title": "ProjeGirisDogrulaLDAPRequest"
    },
    "ProjectUserList": {
      "type": "object",
      "title": "ProjectUserList"
    },
    "TBSAracPlakaRequest": {
      "type": "object",
      "properties": {
        "auth": {
          "$ref": "#/definitions/KullaniciBilgi"
        },
        "plaka": {
          "type": "string"
        }
      },
      "title": "TBSAracPlakaRequest"
    },
    "TuzelSahisDTOResponse": {
      "type": "object",
      "properties": {
        "ad": {
          "type": "string"
        },
        "vergiNo": {
          "type": "string"
        }
      },
      "title": "TuzelSahisDTOResponse"
    },
    "View": {
      "type": "object",
      "properties": {
        "contentType": {
          "type": "string"
        }
      },
      "title": "View"
    }
  }
}